feat!: rename the dashboard's product-named identifiers to stormglass
Fifth rename surface, the embedded UI. None of this is in #198's list —
these are product-named rather than hardware-named, and were classified
hit by hit before editing:

  - web/src/api/tempestApi.ts -> stormglassApi.ts (git mv, with its test
    file and the stormglassApiSource binding). It is the client for THIS
    server's /api routes, not for WeatherFlow's API, so the old name
    pointed at the wrong system.
  - the 'tempest-prefs' localStorage key -> 'stormglass-prefs'
  - package.json's "tempest-app" -> "stormglass-app", and the two matching
    names in package-lock.json, which must agree or `npm ci` fails
  - the page <title> and the two CSS banner comments: "Tempest Weather
    Display" -> "Stormglass"

The prefs key is a deliberate, visible break: each viewer's unit
preferences reset once on first load after upgrade. That is the §2.1
no-migration ruling applied consistently rather than an oversight.

Test-first: the two importing test files were pointed at the new module
name before the source moved, and vitest failed on unresolved imports for
both. 26 files / 183 tests green after, versus 24 / 156 while the two
files could not load.

Five Tempest references in web/src/ are deliberate survivors — the
hardware, not the product: the WeatherFlow data-type header, the LTO
battery voltage comment, the strike-distance comment, and the
'Tempest Station' fallback in Header.tsx with its test. That last one
labels the user's device when STATION_NAME is unset; the station is still
a Tempest, so renaming it would state something false.
