+++ web/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "tempest-app",
+  "name": "stormglass-app",
   "private": true,
   "version": "0.0.0",
   "type": "module",
