+++ web/src/App.css
@@ -1,5 +1,5 @@
 /* ============================================================
-   Tempest Weather Display — Component Styles
+   Stormglass — Component Styles
    ============================================================ */
 
 /* --- Animated background orbs -------------------------------- */
