+++ web/src/index.css
@@ -1,5 +1,5 @@
 /* ============================================================
-   Tempest Weather Display — Base Styles
+   Stormglass — Base Styles
    ============================================================ */
 
 *, *::before, *::after {
